bootstrap is stupid
attempted to change my code with the boot strap bullshit, but all it did was mess everything up and if someone wanted me to work with this they should have taught me it in the beginning instead of waiting till the end when we have no real way of learning it and its pointless anyway. I like how my shit looks the way i designed it so bootstrap is pointless and unnecessary.